Rename `ClaimStakeReward` actions
