+++ Lib9c/Action/ClaimStakeReward.cs
@@ -3,6 +3,7 @@ using Bencodex.Types;
 using Libplanet;
 using Libplanet.Action;
 using Nekoyume.Extensions;
+using Nekoyume.Helper;
 using Nekoyume.Model.Item;
 using Nekoyume.Model.State;
 using Nekoyume.TableData;
@@ -16,8 +17,7 @@ namespace Nekoyume.Action
     [ActionType(ActionTypeText)]
     public class ClaimStakeReward : GameAction
     {
-        public const long ObsoletedIndex = 5_599_601L;
-        private const string ActionTypeText = "claim_stake_reward2";
+        private const string ActionTypeText = "claim_stake_reward3";
 
         internal Address AvatarAddress { get; private set; }
 
@@ -38,7 +38,7 @@ namespace Nekoyume.Action
             }
 
             var states = context.PreviousStates;
-            CheckObsolete(ClaimStakeReward.ObsoletedIndex, context);
+            CheckActionAvailable(ClaimStakeReward2.ObsoletedIndex, context);
             var addressesHex = GetSignerAndOtherAddressesHex(context, AvatarAddress);
             if (!states.TryGetStakeState(context.Signer, out StakeState stakeState))
             {
@@ -89,18 +89,34 @@ namespace Nekoyume.Action
             var accumulatedRewards = stakeState.CalculateAccumulatedRewards(context.BlockIndex);
             foreach (var reward in rewards)
             {
-                var (quantity, _) = stakedAmount.DivRem(currency * reward.Rate);
-                if (quantity < 1 || reward.Type != StakeRegularRewardSheet.StakeRewardType.Item)
+                switch (reward.Type)
                 {
-                    // If the quantity is zero, it doesn't add the item into inventory.
-                    continue;
+                    case StakeRegularRewardSheet.StakeRewardType.Item:
+                        var (quantity, _) = stakedAmount.DivRem(currency * reward.Rate);
+                        if (quantity < 1)
+                        {
+                            // If the quantity is zero, it doesn't add the item into inventory.
+                            continue;
+                        }
+                        ItemSheet.Row row = itemSheet[reward.ItemId];
+                        ItemBase item = row is MaterialItemSheet.Row materialRow
+                            ? ItemFactory.CreateTradableMaterial(materialRow)
+                            : ItemFactory.CreateItem(row, context.Random);
+                        avatarState.inventory.AddItem(item, (int)quantity * accumulatedRewards);
+                        break;
+                    case StakeRegularRewardSheet.StakeRewardType.Rune:
+                        var accumulatedRuneRewards =
+                            stakeState.CalculateAccumulatedRuneRewards(context.BlockIndex);
+                        var runeReward = accumulatedRuneRewards * RuneHelper.CalculateStakeReward(stakedAmount, reward.Rate);
+                        if (runeReward < 1 * RuneHelper.StakeRune)
+                        {
+                            continue;
+                        }
+                        states = states.MintAsset(AvatarAddress, runeReward);
+                        break;
+                    default:
+                        break;
                 }
-
-                ItemSheet.Row row = itemSheet[reward.ItemId];
-                ItemBase item = row is MaterialItemSheet.Row materialRow
-                    ? ItemFactory.CreateTradableMaterial(materialRow)
-                    : ItemFactory.CreateItem(row, context.Random);
-                avatarState.inventory.AddItem(item, (int)quantity * accumulatedRewards);
             }
 
             if (states.TryGetSheet<StakeRegularFixedRewardSheet>(
